Relocate shared lowering types to compiler::lossiness.
Lossiness, LossyField, and LoweredOutput are shared by agent and skill
lowerers; defining them in agents::lower inverted the dependency.
